+++ tools/mission/src/components/citizen.rs
@@ -4,17 +4,17 @@ use simula_video::VideoMaterial;
 use crate::{common::Robot, spawn_robot_with_wallet};
 
 pub const CITIZEN_STARTING_MONEY: u64 = 0;
-pub const CITIZEN_STARTING_ENERGY: u64 = 500;
+pub const CITIZEN_STARTING_ENERGY: f32 = 500.0;
 
 #[derive(Component, Debug, Default, Clone, Copy)]
 pub struct Citizen {
-    energy: u64,
+    energy: f32,
     money: u64,
     follow_ui: Option<Entity>,
 }
 
 impl Robot for Citizen {
-    fn get_energy(&self) -> u64 {
+    fn get_energy(&self) -> f32 {
         self.energy
     }
 
@@ -26,7 +26,7 @@ impl Robot for Citizen {
         self.follow_ui
     }
 
-    fn set_energy(&mut self, energy: u64) {
+    fn set_energy(&mut self, energy: f32) {
         self.energy = energy;
     }
 
@@ -38,7 +38,7 @@ impl Robot for Citizen {
         self.follow_ui = Some(entity);
     }
 
-    fn starting_energy(&self) -> u64 {
+    fn starting_energy(&self) -> f32 {
         CITIZEN_STARTING_ENERGY
     }
 }
