+++ tools/mission/src/ui.rs
@@ -27,7 +27,7 @@ pub fn follow_ui<T>(
                 .title_bar(false)
                 .show(egui_context.ctx_mut(), |ui| {
                     ui.label(name.to_string());
-                    ui.label(format!("Energy: {}", robot.get_energy()));
+                    ui.label(format!("Energy: {}", robot.get_energy() as u32));
                     ui.label(format!("Money: {}", robot.get_money()));
                 });
         }
